Change PasswordJuge to PasswordValidator
Note: PasswordValidator needs refactoring.
